Mobile: Fixes #11123: Fix note listed as "Untitled" after moving to the current folder
Previously, moving a note on mobile could save a note that lacked
properties required for the UI to work (`title`). This pull request adds
`title` to the list of properties that cause a note to be reloaded from
the database before using it to update the note list.

This commit still needs automated tests.