Mobile: Fixes #11276: Fix new note button pushed offscreen on certain Android devices
This commit adjusts an accessibility bug workaround from
(https://github.com/laurent22/joplin/pull/10123), which was first present in Android v3.0.1.

**Background**:
React Native Paper's FAB.Group by default renders an invisible <View> that covers most/all
of Joplin's UI when the FAB.Group is collapsed. This breaks TalkBack on Android's "tap to
navigate" feature, which significantly impacts users that rely on TalkBack's ability to use
the app.
Starting from Joplin v3.0.1, the <View> added by React Native is made smaller by Joplin by
adding margins above and to the left. These margins are determined by the screen size and an
offset. This solution isn't perfect because it means that a small portion of the screen still
blocks TalkBack's tap-to-navigate. As reported in #11276, this workaround also pushed the new
note/new to-do buttons offscren.

**The change**
This commit changes how the FAB.Group's invisible <View> is resized. Previously it was done
with margins. Now, it's done with the top/left absolute positioning properties.

**Note**
If [this upstream pull request](https://github.com/callstack/react-native-paper/pull/4514) is
merged, it should be possible to remove the workaround entirely.

**Remaining steps**
More testing is required before a pull request can be opened. For now, (in part because this is
not a regression from v3.0), I plan to target release-3.2, rather than
release-3.1, to avoid introducing additional regressions in the 3.1 version of the app.

Should fix #11276, #11315.