Desktop,Mobile: Fixes #11335: Auto-match backtick brackets
This commit auto-matches ` characters. Previously, `closeBrackets` was
specified incorrectly (wrong type), which seems to have prevented
bracket matching from taking place.

Note: While this behavior change is a regression on desktop, it isn't on
mobile. If this change is backported to release-3.1, it may make sense
to not apply this change to mobile.