Desktop: Fixes #11594: Fix syncLockGoneError on sync with certain share config
Note: I was able to reproduce the original issue with the following
steps:
1. Create two users on the same team (team user A, team user B).
2. Enable end-to-end encryption for both users.
3. Share a notebook in read-only mode from user A to user B.
4. Accept the share and sync.
5. Share a notebook in writable mode from user B to user A.
6. Accept the share and sync.
7. As user A, make a change to a note from the notebook shared in step 5.
   - Specifically, mark a previously complete task as incomplete.
8. Sync.
9. Observe the "syncLockGone" error.

While this commit seems to fix the error, all changes made from user A
to the modifiable notebook from step 5 are marked as conflicts and
unapplied on sync. As such, I suspect that there is also a bug related
to local (app not server) handling of sync permissions in the 3.2 release.