Adjust styles of edge menu
This commit pushes the menu to the right on large screens. This makes it
appear in the same place as the '+' button. This minimizes the distance
between the initial click and the next click for creating a button.

See also: [Fitts' Law](https://en.wikipedia.org/wiki/Fitts%27s_law)