iOS: Fixes #11711: Fix Markdown toolbar partially covered by keyboard on iOS
This commit replaces a hardcoded autoCompletionBarExtraHeight with a value
determined by react-native-safearea-context. On iOS, when a keyboard is visible,
the extra padding required to show the toolbar on top of the keyboard seems to
be the same as the safearea padding at the top of the screen.

Tested on the following iOS simulators by showing the software keyboard (docked) while in the editor:
- iPhone Xs
- iPhone SE (iOS 18.2)
- iPhone 16 Pro (iOS 18.2)
- iPad Mini (A17 Pro, iOS 18.2)

Tested on the following simulators by enabling the floating keyboard while in the editor:
- iPad Mini (A17 Pro, iOS 18.2)