Improve how editor plugin visibility is saved/restored
Previously, editor plugins visibility was saved based on the view ID of
the plugin. This, however, was problematic because the view ID is based
on the window ID. As a result, editor plugins in secondary windows were
initially shown/hidden in a way that might be unexpected.

With this change, editor plugin visibility is based on the view type ID.
This type ID is the same as the view ID, but excludes the part that's
determined by the window ID. As a result, editor plugins are
shown/hidden based on the last change to that plugin's visibility in any
window.