iOS: Fix share extension
Fixes the following error:
```
-[_EXSinkLoadOperator loadItemForTypeIdentifier:completionHandler:expectedValueClass:options:] nil expectedValueClass allowing {(
    NSUUID,
    NSMutableArray,
    NSValue,
    NSURL,
    NSMutableData,
    NSString,
    NSDictionary,
    NSError,
    NSNumber,
    NSArray,
    NSMutableDictionary,
    NSMutableString,
    UIImage,
    CKShare,
    NSDate,
    _EXItemProviderSandboxedResource,
    NSData
)}
BUG IN CLIENT OF UIKIT: The caller of UIApplication.openURL(_:) needs to migrate to the non-deprecated UIApplication.open(_:options:completionHandler:). Force returning false (NO).
```

Tested on an iOS 18.1 simulator by selecting the "ShareExtension" option in the run scheme dropdown and the "Photos" app as the sharer. Then:
1. From the photos app, selecting and sharing an image.
2. Verifying that this opens Joplin.
   - In the resize image dialog, I chose to resize.
3. Verifying that this attaches the image to a note.