Android: Voice typing: Fix memory leak
Fixes a potentially severe memory leak caused by a mismatched new/free.

Note: Is it feasible to migrate this logic to smart pointers (or
something similar?). WhisperSession is intended to simplify memory
management, but whisperWrapper.cpp still has a number of manual
allocations/frees (most related to JNI, but not all).