WIP: Show resources placeholders in the editor
This change is not a complete fix for rendering image placeholders
within the editor. Currently, the editor uses <p> tags to create
paragraphs. However, <p> elements [cannot contain <div>s and the
<div>s are extracted to
toplevel](https://html.spec.whatwg.org/multipage/grouping-content.html#the-p-element).

This breaks the save and restore functionality, sometimes causing inline
images to become duplicated on restore. As such, this change is
currently a work-in-progress.