Add another (initially disabled) webcrypto-based encryption option
During pull request review, it was suggested that an RSA-OAEP-based
encryption option be provided that uses a length-2048 key (due to the
performance overhead of generating length-4096 keys). This commit adds
such an encryption scheme as "RsaV2". The old RsaV2 has been renamed to
"RsaV3".

As before, migrations to both RsaV2 and RsaV3 are disabled.

Possible concerns:
- See comment at the top of "LongDataWrapper.ts".
- MDN, referring to the key size: "Some organizations are now recommending that it should be 4096."
  - https://developer.mozilla.org/en-US/docs/Web/API/RsaHashedKeyGenParams#moduluslength