Android: Fixes #13193: Fix Markdown editor toolbar buttons sometimes don't work
This change fixes #13193. Previously, the Markdown editor setup script would run multiple
times on Android devices. If the second time the setup script ran was **after** the first
editor was initialized, then the editor reference was cleared and the editor command handler
was unable to call supportsCommand.

Interestingly, supportsCommand did throw an `Error` in this case. However, it was not included
in Joplin's logfile (or logged to the console).