Desktop: Fixes #13561: Don't initially maximize the window on Linux
This pull request works around what seems to be a bug in recent versions
of KDE Plasma: Setting the initial window state to maximized with Plasma
6.5.0 and 6.5.1 and Joplin running in XWayland results in the window not
showing up.
A user has reported a similar issue with Linux Mint + GNOME (I think
GNOME Classic?), so it's possible that this is an issue with a library
used by both KDE Plasma and the version of GNOME in #13600.