Mobile: Fix voice typing feature flag incorrectly shown
The voice typing feature flag's key began with "featureFlag.", which
caused display logic to run that ignored the "public" field. As a quick
workaround, this change updates the name of the voice typing setting to
avoid this rendering logic. The 'showVoiceTypingSettings' utility
function has also been updated to avoid showing voice typing specific
settings when the feature has been disabled.